requirements.txt
django-ninja
ipdb
black==22.3.0
pytest
psycopg2-binary
python-pptx
